requirements.txt
Flask==3.0.1
google-cloud-storage==1.26.0
